requirements.txt
bestconfig==1.3.6
elasticsearch==8.17.1
elasticsearch8-dsl==8.12.0
mysql-connector-python==9.2.0
